+++ spec/dummy/db/migrate/20181021065112_add_country_specifier_proc_number_to_phone.rb
@@ -1,0 +1,5 @@
+class AddCountrySpecifierProcNumberToPhone < ActiveRecord::Migration
+  def change
+    add_column :phones, :country_specifier_proc_number, :string
+  end
+end